fix(workspace): resolve architecture test failures and integration test 500s
- Move providers logic from controller to WorkspaceQueryService (fixes
  controllersAreThin: max 5 constructor params)
- Add @PreAuthorize("permitAll()") to public endpoints (fixes
  controllerMethodsHaveSecurityAnnotations)
- Add IdentityProviderController to workspace context test exclusions
  (fixes dataEndpointsReceiveWorkspaceContext)
- Make ensureAuthenticatedGitLabUser gracefully handle missing gitlab_id
  JWT claim instead of throwing (fixes integration test 500s for users
  authenticated via GitHub)
