+++ server/application-server/src/main/java/de/tum/in/www1/hephaestus/account/IdentityProviderController.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -24,6 +25,7 @@ public class IdentityProviderController {
     }
 
     @GetMapping("/identity-providers")
+    @PreAuthorize("permitAll()")
     @Operation(
         summary = "List available identity providers",
         description = "Returns all enabled identity providers that can be used for login. Public endpoint — no authentication required."
