+++ server/application-server/src/test/java/de/tum/in/www1/hephaestus/architecture/MultiTenancyArchitectureTest.java
@@ -897,7 +897,11 @@ class MultiTenancyArchitectureTest extends HephaestusArchitectureTest {
 
                     // Skip user account operations - these are USER-scoped, not WORKSPACE-scoped.
                     // Users can access their account settings regardless of workspace context.
-                    if (controllerName.contains("Account") || controllerName.contains("FeatureFlag")) {
+                    if (
+                        controllerName.contains("Account") ||
+                        controllerName.contains("FeatureFlag") ||
+                        controllerName.contains("IdentityProvider")
+                    ) {
                         return;
                     }
